[fix] auth and search controller tests
