+++ src/test/java/com/lgtm/easymoney/controllers/FriendControllerTest.java
@@ -15,6 +15,8 @@ import com.lgtm.easymoney.models.User;
 import com.lgtm.easymoney.payload.req.FriendshipReq;
 import com.lgtm.easymoney.payload.rsp.ProfileRsp;
 import com.lgtm.easymoney.payload.rsp.ProfilesRsp;
+import com.lgtm.easymoney.security.JwtAuthenticationEntryPoint;
+import com.lgtm.easymoney.security.JwtTokenProvider;
 import com.lgtm.easymoney.services.FriendService;
 import com.lgtm.easymoney.services.impl.UserServiceImpl;
 import java.util.ArrayList;
@@ -24,8 +26,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
@@ -38,8 +39,7 @@ import org.springframework.test.web.servlet.ResultActions;
  * */
 
 @RunWith(SpringRunner.class)
-@SpringBootTest
-@AutoConfigureMockMvc
+@WebMvcTest(FriendController.class)
 public class FriendControllerTest {
   @Autowired
   private MockMvc mvc;
@@ -47,8 +47,13 @@ public class FriendControllerTest {
   @MockBean
   private FriendService friendService;
 
+  // We test jwt functionalities in integration tests instead
   @MockBean
   private UserServiceImpl userService;
+  @MockBean
+  private JwtAuthenticationEntryPoint jwtAuthenticationEntryPoint;
+  @MockBean
+  private JwtTokenProvider jwtTokenProvider;
 
   private FriendshipReq friendshipReq;
   private ProfilesRsp profilesRsp;
@@ -72,15 +77,6 @@ public class FriendControllerTest {
     List<ProfileRsp> res = new ArrayList<>();
     res.add(profileRsp);
     profilesRsp = new ProfilesRsp(res);
-  }
-
-  @Test
-  public void noAccessForUnauthenticatedUsers() throws Exception {
-    mvc.perform(post("/friend/add")).andExpect(status().isUnauthorized());
-    mvc.perform(put("/friend/accept")).andExpect(status().isUnauthorized());
-    mvc.perform(delete("/friend/{uid}", String.valueOf(uid2))).andExpect(status().isUnauthorized());
-    mvc.perform(get("/friend")).andExpect(status().isUnauthorized());
-    mvc.perform(get("/friend/pending")).andExpect(status().isUnauthorized());
   }
 
   @Test
