+++ src/test/java/com/lgtm/easymoney/controllers/SearchControllerTest.java
@@ -9,9 +9,13 @@ import com.lgtm.easymoney.enums.UserType;
 import com.lgtm.easymoney.exceptions.ResourceNotFoundException;
 import com.lgtm.easymoney.payload.rsp.ProfileRsp;
 import com.lgtm.easymoney.payload.rsp.ProfilesRsp;
+import com.lgtm.easymoney.security.JwtAuthenticationEntryPoint;
+import com.lgtm.easymoney.security.JwtTokenProvider;
 import com.lgtm.easymoney.services.SearchService;
 import java.util.ArrayList;
 import java.util.List;
+
+import com.lgtm.easymoney.services.impl.UserServiceImpl;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -34,6 +38,15 @@ public class SearchControllerTest {
   private MockMvc mvc;
   @MockBean
   private SearchService searchService;
+
+  // We test jwt functionalities in integration tests instead
+  @MockBean
+  private UserServiceImpl userService;
+  @MockBean
+  private JwtAuthenticationEntryPoint jwtAuthenticationEntryPoint;
+  @MockBean
+  private JwtTokenProvider jwtTokenProvider;
+
   private ProfilesRsp profilesRsp;
   private ProfileRsp profileRsp;
   private List<ProfileRsp> profileList;
